+++ config/parameters.js
@@ -64,6 +64,7 @@ exports = module.exports = Object.assign({}, basicParameters, {
 	fileSystem: {
 		root: process.env.fs_root_path,
 		tmp: process.env.fs_tmp_path || '/tmp',
+		system: process.env.fs_system_path,
 	},
 	video: {
 		socket_root: '/tmp'