+++ src/SystemSettings/FSSystemSettings.ts
@@ -1,0 +1,64 @@
+import * as path from 'path';
+import * as fs from 'fs-extra';
+import ISystemSettings from './ISystemSettings';
+
+const DEFAULT_VOLUME = 100;
+
+export default class FSSystemSettings implements ISystemSettings {
+
+	private settings: {
+		volume?: number;
+	} = {};
+	private loadedFromFS: boolean = false;
+
+	constructor(private fileSystemBasePath: string) {}
+
+	public async getVolume(): Promise<number> {
+		await this.loadFromFileSystemIfNotLoaded();
+		if (typeof this.settings.volume === 'undefined') {
+			return DEFAULT_VOLUME;
+		}
+		return this.settings.volume;
+	}
+
+	public async setVolume(volume: number): Promise<void> {
+		this.settings.volume = volume;
+		await this.saveToFileSystem();
+	}
+
+	private async loadFromFileSystemIfNotLoaded() {
+		if (!this.loadedFromFS) {
+			await this.loadFromFileSystem();
+			this.loadedFromFS = true;
+		}
+	}
+
+	private async loadFromFileSystem() {
+		const filePath = this.getSettingsFilePath();
+		try {
+			if (await fs.pathExists(filePath)) {
+				const fileContentsBuffer = await fs.readFile(filePath);
+				this.settings = JSON.parse(fileContentsBuffer.toString());
+			} else {
+				this.settings = {};
+			}
+		} catch (error) {
+			console.error('Failed to load system settings from file system', error);
+			this.settings = {};
+		}
+	}
+
+	private async saveToFileSystem() {
+		const filePath = this.getSettingsFilePath();
+		const settingsJSON = JSON.stringify(this.settings);
+		try {
+			await fs.writeFile(filePath, settingsJSON);
+		} catch (error) {
+			console.error('Failed to save system settings to file system', error);
+		}
+	}
+
+	private getSettingsFilePath() {
+		return path.join(this.fileSystemBasePath, 'system_settings.json');
+	}
+}