+++ src/SystemSettings/ISystemSettings.ts
@@ -1,0 +1,4 @@
+export default interface ISystemSettings {
+	getVolume(): Promise<number>;
+	setVolume(volume: number): Promise<void>;
+}